Remove the "first use generates config file" magic
I think its cleaner if just exit with an error if emscripten is not yet
configured.  Especially since we already have an error mesasge for
"config file not found" so this change means that first time users will
now see that message (in which we now recommend the use of
`--generate-config`).

There are couple of reasons why I think this is a good idea:

1. We already have a error message for missing config file
2. Currently when the config file is generated we return 0 without
   actually doing anything, which could confuse build systems.
3. In the case of a misconfigured systems (e.g. an accidentailly
   removed config file, or a EM_CONFIG pointing the wrong place)
   the solution is probably not to simply just create a new
   config, but to inform the developer and let them take action.